pyassem: Rework block ordering algorithm.
The original code from Python2 stdlib had over-engineered algorithm with
"dominators" (real or not), which actually had to be workaround already
to get order compatible with CPython3.5, and which produced rather erratic
orderings for more complex constructs.

Instead, following idea is applied: The input is a structured program with
a structured CFG. During compilation, we don't apply any advanced CFG
transformations. Then, the basic blocks should be output in the same
order as they are processed by compiler (based on FlowGraph.nextBlock()
calls).

This was tested to work as expected (i.e. produce the same output as
CPython3.5) on the package source files themselves. This change removes
the need for quite a lot code in pyassem.py, but it's left there just
in case, until more global testing is done.