Remove NestedScopeMixin class
Summary:
This is the start of cleaning up how the compiler handles code generation across different target objects (classes, generators, functions, lambda, etc...).  Currently the compiler uses inheritance to perform this - but inheritance would be better suited for supporting different code generation features (e.g. Python versions or runtimes).

The first part of this is pretty simple - it just gets rid of the NestedScopeMixin which is an unused extensibility point which allows you to configure what subclass is going to be used for the different target object.

Test Plan: python3 -m test_compiler