Code Python 3.7 compiler - LOAD_METHOD/CALL_METHOD
Summary: First step towards having a compiler that supports Python 3.7 - we now have a code generator which subclasses the normal one, and modifies the code we generate for calling methods.  In order to support running/testing on Python 3.6 we need to define the new opcodes manually.

Test Plan: python3 -m test_compiler