Python 3.7 future import support
Summary:
Adds support for Python 3.7's from __future__ import annotations, and removes support of from __future__ import generator_stop.

annotations is pretty simple from a code generation perspective, we just need to output the annotation as a LOAD_CONST instead of as normal Python objects.  But that requires a bunch of of code to convert expressions back into strings (which is how CPython accomplishes this as well).

Test Plan: python3 -m test_compiler