Cleanup visitor in prep for rewriter
Summary:
The existing visitor is kind of funky...  It has both a "visitor" and a "walker" class.  The walker is actually the class named ASTVisitor, and the visitor is some random other class that doesn't inherit from anything.  When you create the walker class is modifies the visitor object to add a random "visit" method on it which you can use.

This cleans all of that up and does the typical visitor pattern thing - you inherit from the visitor class.

Test Plan: python3 -m test_compiler