Optimize bin op support
Summary:
Continuing on the AST based optimization this adds support for folding binary operations.

The new optimization uses the same logic for limiting the size of objects on which constant folding is performed, but the limits are different than the 3.6 peephole optimizer so we now parameterize those functions with a limits class.

Test Plan: python3 -m test_compiler