Match stack depth calculation
Summary: 3.7 adds a new stack depth calculation algorithm which is more precise when accounting for jumps - it's now account for the depth depending on whether or not the jump is taken.  It also moves to a non-recursive algorithm as well, which we match for making future changes easier.

Test Plan:
python3 -m test_compiler

Test coverage for these specific changes from the the test_sbs_stdlib tests where we logout the stack depth size for all functions in the standard library.