symbols: Basic block-level scoping infrastructure.
We introduce BlockScope, as a subclass of FunctionScope, to piggyback on
*some* of the scoping processing done for FunctionScope (but we otherwise
going to ignore the rest of it).

What we mostly need BlockScope for is like a container for remappings of
var_name -> var_name.NN (numeric counter). We'd like to apply remapping
"on the fly" (vs a separate pass on AST with explicit variable rewriting).
The mapping is stored in Scope.defs (which previously was set emulated
using dict, i.e. contained 1 as a value).

For the rest of functionality, we try to forward it to the containing
function scope (and for this, we maintain .func_parent pointer in BlockScope,
note that it maybe also module or class scope).