symbols, pycodegen: Prototype of block-level scope for "for" variables.
Due to the fact that this package doesn't include tokenizer and AST parser,
this prototype applies block-level scope to all "for" statements, whereas
the original proposal is to apply it only to the loops of new "for const"
syntax. This prototype also explores block-level scope aspect, not the
constness aspect.

The implementation is based on basic block-level scoping support introduced
in the previous commit.

The only interesting part here is that we actually have to walk BlockScope
subtree to (potentially) remap block-scope variable names. The original
inspriring idea was to avoid scope-iteration at all, but we clearly don't
do enough processing in symbols.py for that. And note that to achieve that,
we'd need to duplicate mapping information in different scopes, i.e. would
achieve O(1) lookups, but at the price of additional memory usage (likely,
quadratic). OTOH, due to processing already done by symbols.py, we already
have sparse namespacing info (comparing to the original AST), and in most
cases, the scope tree will be rather shallow (or put it differently, price
of O(n) compile-time lookups will be paid only by the code which (ab)uses
nested scopes too much).